+++ fs_util/src/lib.rs
@@ -1,0 +1,7 @@
+mod safe_move;
+mod same_content;
+mod walk_dir;
+
+pub use safe_move::safe_move;
+pub use same_content::same_content;
+pub use walk_dir::{DirEntry, DirEntryError, WalkDir};