+++ service/src/state.rs
@@ -18,7 +18,6 @@ use crate::{
     repository::{self, RepositoryHandle, RepositoryHolder, RepositorySet},
     tls::TlsConfig,
     transport::remote::{AcceptedRemoteConnection, RemoteClient, RemoteServer},
-    utils,
 };
 use ouisync::{
     crypto::{cipher::SecretKey, Password, PasswordSalt},
@@ -1809,16 +1808,16 @@ impl State {
     // Find all repositories in the store dir and open them.
     async fn load_repositories(&self) {
         let Some(store_dir) = self.store.get() else {
-            tracing::warn!("store dir not specified");
+            tracing::debug!("store dir not specified");
             return;
         };
 
         if !fs::try_exists(&store_dir).await.unwrap_or(false) {
-            tracing::error!("store dir doesn't exist");
+            tracing::debug!("store dir doesn't exist");
             return;
         }
 
-        let mut walkdir = utils::walk_dir(store_dir);
+        let mut walkdir = fs_util::WalkDir::new(store_dir).into_stream();
 
         while let Some(entry) = walkdir.next().await {
             let entry = match entry {